requirements.txt
matplotlib==3.1.1
numpy==1.17.3
numpy==1.17.2
pandas==1.0.3
pandas==0.25.1
Pillow==6.2.1
Pillow==6.2.0
Pillow==10.2.0
pytorch_msssim==1.0.0
PyYAML==5.4.1
PyYAML==6.0.1
torch==2.2.1
torchvision==0.17.1
